requirements.txt
PyQt6>=6.6.0
PyYAML>=6.0.0
requests>=2.31.0
pypinyin>=0.50.0
httpx>=0.26.0
